+++ webrtc-android-sample-app/src/main/AndroidManifest.xml
@@ -98,10 +98,6 @@
             android:exported="true"
             android:theme="@style/Theme.AppCompat.DayNight"
             android:configChanges="orientation|keyboard|screenSize|smallestScreenSize|screenLayout"/>
-        <activity android:name=".MP3ConvertActivity"
-            android:exported="true"
-            android:theme="@style/Theme.AppCompat.DayNight"
-            android:configChanges="orientation|keyboard|screenSize|smallestScreenSize|screenLayout"/>
     </application>
 
 </manifest>
